Code refactoring and bug fixed
