+++ src/hooks/useInteractive.ts
@@ -150,6 +150,8 @@ export const useInteractive = (
         event.stopPropagation()
         if (currentSelectedId !== component.id) {
           dispatch.components.select(component.id)
+          dispatch.text.reset()
+          dispatch.text.setTextValue(ref.current?.innerHTML || '')
         }
       },
       componentId: component.id,
